Require files in gemspec via git, upgrade Rake dep.
The minimum rake requirement should be at least version 10.0.0 as
there are some security issues associated with the earlier versions.

Nobody has time to list files by hand in the gemspec. I'm pretty
sure there is an open PR about this, that just goes and modifies
the list by hand.